ENH: Allow to set group / mode / owner of newly created files
As the location usually is not accessible by our user, we likely also want any created files to be owned by somebody else.
With a ** glob and an associative array, we can implement a simple file delta, and then apply any passed file modifications to the newly created files (without affecting existing ones).