ENH: Add withTempdir variant of withTempfile that delegates to the latter with a new hidden --directory flag
The tempdir command isn't that good for scripting use cases where the CWD needs to be maintained. In order to make withTempfile handle directories, we just need to slightly adapt the tempfile generation and do recursive cleanup, so it's better to delegate the functionality to it, but offer a separate withTempdir command for a cleaner API and precise help.