processPassed{Month,Year}s: BUG: Also need to create the queried key if it doesn't exist
I didn't receive a monthly window-title report for Hannah, just for me. That's because on findus, reports are configured both for me and Hannah. But processPassedMonths only created the database for the first query (exit status 1) for my user, but then simply skipped the creation of the entry for Hannah, as the database then existed. But without such an entry, on the following month Hannah again will be skipped as there's again no stored (previous) month.