Revert "each{Dir,File}: ENH: Also accept -- before SIMPLECOMMAND"
Not a good idea, because in order to pass non-existing FILE(s) when -c|--command is used one now has to pass -- -- FILE [...], which is strange. And even that case isn't handled properly, because the addition of SIMPLECOMMAND doesn't check for an empty @cmd and mistakenly appends a quoted ''!
I'll handle this use case by allowing --exec instead.

This reverts commit 3b8f917cfa91179e165b26e4cc4585053dad863f.