ENH: Allow generation without passed FILE(s) via --generation-{command,exec}
The supplied command may also / only be able to provide the filespecs without any inputs. If the inputs are hard-coded, it's easier to inline those directly in the SELECT-COMMAND than passing them separately as FILE(s) (even if --files makes that easier for wrappers).
It's a bit unfortunate that for pure generator use the withSelectedFiles name is misleading, but both code reuse and the possibility to combine generator and select-command make it unattractive to implement a separate withGeneratedFiles command instead.