Add withRestoredFile variant of withRemovedFile
Restore is just a variation that does not remove the source files, only saves and finally restores them.
To sidestep the merge problem with directories, only files are supported so far.
The overwrite could be skipped if the file wasn't actually touched, but checking with cmp alone would be more effort and ignore file ownership issues. Rather assume that those file(s) that are passed are likely be overwritten or that the superfluous restore just doesn't matter.