Refactoring: Unify withNewestFile[s] into the singular form
That there is newestFile[s] is an implementation detail that (now that the behavior especially in regards to recursion has been aligned) can already be hidden.
withNewestFile has to parse the newestFiles arguments, anyway, so we can decide which one to use; this now is purely an internal optimization, not something the user has to be bothered with.
With that, all the plural variants of withNewest* and open-newest-* can be removed as well; a huge win.