Refactoring: Replace array of possibly interpolated arg with quotedArg
I needed the array when making a distinction with $isNeedEval. Now the result is eval'd all the time; using the quoted array looks weird and in a few scripts I actually forgot the single-quoting (which caused latent bugs with empty arguments or those that require quoting).
Directly using (and clearing if already interpolated) the quotedArg also improves tracing the command execution, as I now see the actual arguments instead of the quoted array variable. It might be a bit more inefficient, though, as the whole array needs to be quoted and interpreted (but most commands just support a single argument).