process*: ENH: Add --status-99-is-noop flag
The processing command may indicate no action with a status 99. In that case, we still want to update the database / delete empty files.
We could put an "acceptStatus 0 99" in between, but that would drop the processing vs. noop information.
The new flag changes the status calculation so that a 99 is not treated as command failure.

System-info: Ubuntu 24.04.3 LTS, x86_64
Platform-info: bash 5.2.21(1)-release, Bats 1.10.0